+++ package.json
@@ -68,7 +68,6 @@
         "reselect": "^4.0.0",
         "styled-components": "^5.0.0",
         "styled-system": "^5.1.4",
-        "styles": "^0.2.1",
         "wait-on": "^3.3.0",
         "yup": "^0.28.1"
     },
