+++ src/admin/baseComponents/button/OFButton.stories.js
@@ -3,7 +3,7 @@ import React from 'react'
 
 export default {
     component: OFButton,
-    title: 'Admin/Components/Button',
+    title: 'Admin|Button',
 }
 
 export const defaultUsage = () => <OFButton>Click me</OFButton>
