+++ src/admin/baseComponents/form/radiobutton2/OFRadioButton.stories.js
@@ -6,7 +6,7 @@ import { Field, Formik } from 'formik'
 
 export default {
     component: OFRadioButton,
-    title: 'Admin/Components/RadioButton',
+    title: 'Admin|RadioButton',
 }
 
 const name = 'picker'
