+++ src/admin/baseComponents/form/switch/OFSwitch.stories.js
@@ -1,4 +1,4 @@
-import OFFormControl from '../formcontrol/OFFormControl'
+import OFFormControl from '../formcontrol2/OFFormControl'
 import React from 'react'
 import { Field, Formik } from 'formik'
 import FormControlLabel from '@material-ui/core/FormControlLabel'
@@ -6,7 +6,7 @@ import { OFSwitch } from './OFSwitch'
 
 export default {
     component: OFSwitch,
-    title: 'Admin/Components/Switch',
+    title: 'Admin|Switch',
 }
 
 export const defaultUsage = () => (
