+++ src/admin/project/new/Step1.projectIdChange.js
@@ -2,7 +2,7 @@ import React from 'react'
 import Menu from '@material-ui/core/Menu'
 import CardContent from '@material-ui/core/CardContent'
 import TranslatedTypography from '../../baseComponents/TranslatedTypography'
-import OFFormControlInputFormiked from '../../baseComponents/form/formcontrol/OFFormControlInputFormiked'
+import OFFormControlInputFormiked from '../../baseComponents/form/formcontrol2/OFFormControlInputFormiked'
 import Box from '@material-ui/core/Box'
 import OFButton from '../../baseComponents/button/OFButton'
 import { useTranslation } from 'react-i18next'
